Changed in GroupDto StudentList from Student to StudentDto

+++ src/main/java/com/example/demo/student/StudentService.java
@@ -1,8 +1,5 @@
 package com.example.demo.student;
 
-import com.example.demo.group.Group;
-import com.example.demo.group.GroupRepository;
-import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
@@ -14,13 +11,11 @@ import java.util.stream.Collectors;
 public class StudentService {
 
     private final StudentRepository studentRepository;
-    private final GroupRepository groupRepository;
     private final StudentMapper studentMapper;
 
-    public StudentService(StudentRepository studentRepository, StudentMapper studentMapper, GroupRepository groupRepository) {
+    public StudentService(StudentRepository studentRepository, StudentMapper studentMapper) {
         this.studentRepository = studentRepository;
         this.studentMapper = studentMapper;
-        this.groupRepository = groupRepository;
     }
 
     public List<StudentDto> getStudents() {
@@ -31,14 +26,11 @@ public class StudentService {
 
     public StudentDto getStudentById(Long studentId) {
         return studentRepository.findById(studentId)
-                .map(studentMapper::toStudentDto).orElseThrow(() -> new IllegalStateException("No student"));
+                .map(studentMapper::toStudentDto).orElseThrow(() -> new IllegalStateException("No student with id " + studentId));
     }
 
     public void addNewStudent(StudentDto studentDto) {
-        Group group = groupRepository.findById(studentDto.getGroupId()).orElseThrow(
-                () -> new IllegalStateException("group with id " + studentDto.getGroupId() + " does not exists"));
         Student student = studentMapper.toStudent(studentDto);
-        group.addNewStudentToStudentList(student);
         studentRepository.save(student);
 
     }
@@ -46,10 +38,11 @@ public class StudentService {
     public void deleteStudent(Long studentId) {
         Student student = studentRepository.findById(studentId).orElseThrow(
                 () -> new IllegalStateException("student with id " + studentId + " does not exists"));
-        Group group = groupRepository.findById(student.getGroup().getId()).orElseThrow(
-                () -> new IllegalStateException("group with id " + student.getGroup().getId() + " does not exists"));
-        group.removeStudentFromStudentList(student);
-        studentRepository.deleteById(studentId);
+        if (student.getGroup() == null) {
+            studentRepository.deleteById(studentId);
+        } else {
+            throw new IllegalStateException("student " + studentId + " in group with id " + student.getGroup().getId());
+        }
     }
 
     @Transactional
@@ -59,8 +52,5 @@ public class StudentService {
         student.setName(studentDto.getName());
         student.setEmail(studentDto.getEmail());
         student.setDateOfBirth(studentDto.getDateOfBirth());
-        Group group = groupRepository.findById(studentDto.getGroupId()).orElseThrow(
-                () -> new IllegalStateException("group with id " + studentDto.getGroupId() + " does not exists"));
-        group.addNewStudentToStudentList(student);
     }
 }
